requirements.txt
django
authlib
requests
django-storages[s3]
mysqlclient
Pillow
ollama
pytz